Story-Generator-main/requirements.txt
gunicorn
google-generativeai
Flask
requests
Pillow